Add some more content
Computing in HEP had a long history and its own vocabulary. We should
probably take a moment to make sure we speak the same language. Where we
need resouces seperates nicely into near the detector and globally.
Analysis seems to logically fit best at the end, as it combines
simulated and reconstucted data ... to make a scientific statement.