Add section on data and analysis preservation
During the ATLAS analysis interview I noticed that you stumbled across
the question on how one could re-do an analysis. While you discussed two
fair points: reading the code ;-) and the support note there are other
structures to support this. While there is no difinitive solution yet
there are good ideas that may be interesting to talk about. Since this
requires an understanding of what is needed to use the data and perfrom
an analysis the end seems a good place to put this section.